Postgres Support

+++ bin/index.js
@@ -1,16 +1,27 @@
 const Discord = require("discord.js");
 const fs = require("fs");
 const util = require("plugit-util");
+const pg = require('pg');
+
+
+//Create Clients
+const db = new pg.Client({
+  connectionString: process.env.DATABASE_URL,
+  ssl: true,
+});
 const client = new Discord.Client();
 
-var commands = {}; //Create Dictionary to store Commands
+
+//Initialize Variables
+var commands = {};
 var events = {};
+
 
 //Help Command
 commands["help"] = {
   name: "help",
   help: "Display A List of Commands And Their Features!",
-  main: function(bot, msg) {
+  main: function(bot, db, msg) {
     embed = new Discord.RichEmbed()
     .setTitle("Bot Commands")
     for (command in commands) {
@@ -24,7 +35,8 @@ commands["help"] = {
   }
 }
 
-var loadCommands = function() {
+
+var loadModules = function() {
     //Load NPM Modules
     var modules = util.modules.getModules();
     for (let file of modules) {
@@ -32,7 +44,7 @@ var loadCommands = function() {
         var module = require(file.path + "/" + file.pkg.main);
       }
       catch (e) {
-        console.log("Skipped " + module)
+        console.log(e)
       }
       finally {
         for (command in module.commands) {
@@ -49,50 +61,79 @@ var loadCommands = function() {
     var files = fs.readdirSync(__dirname + "/../modules/");
     for (let file of files) {
       if (file.endsWith('.js')) {
-        var module = require(__dirname + "/../modules/" + file);
-        for (command in module.commands) {
-          commands[command] = module.commands[command];
+        try {
+          var module = require(__dirname + "/../modules/" + file);
         }
-        for (event in module.events) {
-          events[event] = module.events[event]
+        catch(e) {
+          console.log(e)
         }
-        if(process.env.debug) {
-            console.log("Loaded " + file.slice(0, -3) + " Module");
+        finally {
+          for (command in module.commands) {
+            commands[command] = module.commands[command];
+          }
+
+          for (event in module.events) {
+            events[event] = module.events[event]
+          }
+
+          if(process.env.debug == true) {
+              console.log("Loaded " + file.slice(0, -3) + " Module");
+          }
         }
       }
     }
-    console.log("———— Modules Loaded! ————");
 }
+
 
 client.on('ready', () => {
   console.log(`Logged in as ${client.user.tag}!`);
   console.log(`Connected to ${client.guilds.size} guilds, serving ${client.users.size} users.`);
-  loadCommands();
-  util.modules.logModules(util.modules.getModules());
   client.user.setGame(process.env.playing)
 
+  //Load commands & events
+  loadModules();
+  util.modules.logModules(util.modules.getModules());
+
+  //Handle events
+  if (events['ready']) {
+    events['ready'].main(client, db)
+  }
+
   for (event in events) {
-    client.on(event, (arg1, arg2, arg3) => {
-      events[event].main(client, arg1)
+    if (event == 'message') return
+    client.on(event, (arg1) => {
+      events[event].main(client, db, arg1)
     })
   }
 });
 
+
+//Handle Commands
 client.on('message', msg => {
-  if ( msg.author.bot ) return
+  if ( msg.author.bot ) return //If message is from a bot ignore.
   if ( msg.content.indexOf(process.env.prefix) !== -1 ) {
     var command = msg.content.split(process.env.prefix)[1].split(" ")[0];
     if(command in commands){
-      if(process.env.delete_commands == true){
-         msg.delete();
-      }
-      commands[command].main(client, msg);
+      if(process.env.delete_commands == true) msg.delete()
+      commands[command].main(client, db, msg);
     }
   }
 
   if (events["message"]) {
-    events["message"].main(client, msg)
+    events["message"].main(client, db, msg)
   }
 });
 
-client.login(process.env.token);
+
+
+console.log("———————— Plugit! ————————");
+
+db.connect() //Connect to database
+  .then(() => console.log(`Successfully Connected To Database`))
+  .catch(e => {
+    console.error('Connection error', e.stack)
+    process.exit()
+  })
+
+
+client.login(process.env.token); //Connect to Discord
